requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
bcrypt>=4.1
django-cors-headers>=4.3
